requirements.txt
absl-py==0.8.0
astor==0.8.0
dlib==19.18.0
gast==0.2.2
google-pasta==0.1.7
grpcio==1.24.0
h5py==2.10.0
imutils==0.5.3
Keras==2.2.4
Keras-Applications==1.0.7
Keras-Preprocessing==1.0.9
Markdown==3.1.1
numpy==1.17.2
opencv-python==4.1.1.26
opt-einsum==3.1.0
protobuf==3.9.2
PyYAML==5.1.2
scipy==1.3.1
six==1.12.0
tensorflow-estimator==1.13.0rc0
tensorflow-gpu==1.13.0rc1
termcolor==1.1.0
Werkzeug==0.16.0
wrapt==1.11.2
